pyproject.toml
[tool.poetry]
name = "matomocamp-live"
version = "0.1.0"
description = ""
authors = ["Lukas Winkler <git@lw1.at>"]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.0.2"
python-dateutil = "^2.8.2"
Babel = "^2.9.1"
libsass = "^0.21.0"
gunicorn = "^20.1.0"
setproctitle = "^1.2.2"
requests = "^2.28.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
